Create Cross Validation and Feature Importances
This file is run on a subset of the full data which contains all instances that had non-zero transaction values (the target variable). Three algorithms were run, decision tree regressor, random forest regressor and adaboost regressor. After this, the importance of each feature as returned by the models was plotted. To further verify the results, k-fold cross validation was applied with k set to 5.